+++ e2e-tests/development-runtime/cypress/integration/navigation/linking.js
@@ -110,7 +110,7 @@ describe(`navigation`, () => {
 
   describe(`Supports unicode characters in urls`, () => {
     it(`Can navigate directly`, () => {
-      cy.visit(`/안녕`).waitForRouteChange()
+      cy.visit(encodeURI(`/안녕`)).waitForRouteChange()
       cy.getTestElement(`page-2-message`)
         .invoke(`text`)
         .should(`equal`, `Hi from the second page`)
@@ -126,7 +126,7 @@ describe(`navigation`, () => {
     })
 
     it(`should show 404 page when url with unicode characters point to a non-existent page route when navigating directly`, () => {
-      cy.visit(`/안녕404/`, {
+      cy.visit(encodeURI(`/안녕404/`), {
         failOnStatusCode: false,
       }).waitForRouteChange()
 
@@ -143,23 +143,76 @@ describe(`navigation`, () => {
     })
   })
 
-  describe(`All location changes should trigger an effect`, () => {
-    beforeEach(() => {
-      cy.visit(`/navigation-effects`).waitForRouteChange()
-    })
+  if (Cypress.env("HOT_LOADER") !== `fast-refresh`) {
+    describe(`All location changes should trigger an effect (react-hot-loader)`, () => {
+      beforeEach(() => {
+        cy.visit(`/navigation-effects`).waitForRouteChange()
+      })
 
-    it(`should trigger an effect after a search param has changed`, () => {
-      cy.getTestElement(`effect-message`).invoke(`text`).should(`eq`, `Waiting for effect`)
-      cy.getTestElement(`send-search-message`).click().waitForRouteChange()
-      cy.getTestElement(`effect-message`).invoke(`text`).should(`eq`, `?message=searchParam`)
-    })
+      it(`should trigger an effect after a search param has changed`, () => {
+        cy.findByTestId(`effect-message`).should(
+          `have.text`,
+          `Waiting for effect`
+        )
+        cy.findByTestId(`send-search-message`).click().waitForRouteChange()
+        cy.findByTestId(`effect-message`).should(
+          `have.text`,
+          `?message=searchParam`
+        )
+      })
 
-    it(`should trigger an effect after the hash has changed`, () => {
-      cy.getTestElement(`effect-message`).invoke(`text`).should(`eq`, `Waiting for effect`)
-      cy.getTestElement(`send-hash-message`).click().waitForRouteChange()
-      cy.getTestElement(`effect-message`).invoke(`text`).should(`eq`, `#message-hash`)
+      it(`should trigger an effect after the hash has changed`, () => {
+        cy.findByTestId(`effect-message`).should(
+          `have.text`,
+          `Waiting for effect`
+        )
+        cy.findByTestId(`send-hash-message`).click().waitForRouteChange()
+        cy.findByTestId(`effect-message`).should(`have.text`, `#message-hash`)
+      })
+
+      it(`should trigger an effect after the state has changed`, () => {
+        cy.findByTestId(`effect-message`).should(`have.text`, ``)
+        cy.findByTestId(`send-state-message`).click().waitForRouteChange()
+        cy.findByTestId(`effect-message`).should(
+          `have.text`,
+          `this is a message using the state`
+        )
+      })
     })
-  })
+  }
+
+  // TODO: Check if this is the correct behavior
+  if (Cypress.env("HOT_LOADER") === `fast-refresh`) {
+    describe(`All location changes should trigger an effect (fast-refresh)`, () => {
+      beforeEach(() => {
+        cy.visit(`/navigation-effects`).waitForRouteChange()
+      })
+
+      it(`should trigger an effect after a search param has changed`, () => {
+        cy.findByTestId(`effect-message`).should(`have.text`, ``)
+        cy.findByTestId(`send-search-message`).click().waitForRouteChange()
+        cy.findByTestId(`effect-message`).should(
+          `have.text`,
+          `?message=searchParam`
+        )
+      })
+
+      it(`should trigger an effect after the hash has changed`, () => {
+        cy.findByTestId(`effect-message`).should(`have.text`, ``)
+        cy.findByTestId(`send-hash-message`).click().waitForRouteChange()
+        cy.findByTestId(`effect-message`).should(`have.text`, `#message-hash`)
+      })
+
+      it(`should trigger an effect after the state has changed`, () => {
+        cy.findByTestId(`effect-message`).should(`have.text`, ``)
+        cy.findByTestId(`send-state-message`).click().waitForRouteChange()
+        cy.findByTestId(`effect-message`).should(
+          `have.text`,
+          `this is a message using the state`
+        )
+      })
+    })
+  }
 
   describe(`Route lifecycle update order`, () => {
     it(`calls onPreRouteUpdate, render and onRouteUpdate the correct amount of times on route change`, () => {