+++ packages/gatsby-plugin-image/src/components/gatsby-image.browser.tsx
@@ -7,6 +7,7 @@ import React, {
   ImgHTMLAttributes,
   useState,
   RefObject,
+  CSSProperties,
 } from "react"
 import {
   getWrapperProps,
@@ -15,22 +16,29 @@ import {
 } from "./hooks"
 import { PlaceholderProps } from "./placeholder"
 import { MainImageProps } from "./main-image"
-import { Layout } from "../utils"
+import { Layout } from "../image-utils"
 
-export type GatsbyImageProps = Omit<
-  ImgHTMLAttributes<HTMLImageElement>,
-  "placeholder" | "onLoad" | "src" | "srcSet" | "width" | "height"
-> & {
+// eslint-disable-next-line @typescript-eslint/interface-name-prefix
+export interface GatsbyImageProps
+  extends Omit<
+    ImgHTMLAttributes<HTMLImageElement>,
+    "placeholder" | "onLoad" | "src" | "srcSet" | "width" | "height"
+  > {
   alt: string
   as?: ElementType
   className?: string
-  image: ISharpGatsbyImageData
+  imgClassName?: string
+  image: IGatsbyImageData
+  imgStyle?: CSSProperties
+  backgroundColor?: string
+  objectFit?: CSSProperties["objectFit"]
+  objectPosition?: CSSProperties["objectPosition"]
   onLoad?: () => void
   onError?: () => void
   onStartLoad?: Function
 }
 
-export interface ISharpGatsbyImageData {
+export interface IGatsbyImageData {
   layout: Layout
   height?: number
   backgroundColor?: string
@@ -51,6 +59,7 @@ export const GatsbyImageHydrator: FunctionComponent<GatsbyImageProps> = function
   onStartLoad,
   image,
   onLoad: customOnLoad,
+  backgroundColor,
   ...props
 }) {
   if (!image) {
@@ -59,7 +68,7 @@ export const GatsbyImageHydrator: FunctionComponent<GatsbyImageProps> = function
     }
     return null
   }
-  const { width, height, layout, images, backgroundColor } = image
+  const { width, height, layout, images } = image
 
   const root = useRef<HTMLElement>()
   const hydrated = useRef(false)
@@ -68,7 +77,7 @@ export const GatsbyImageHydrator: FunctionComponent<GatsbyImageProps> = function
   >(null)
   const lazyHydrator = useRef<(() => void) | null>(null)
   const ref = useRef<HTMLImageElement | undefined>()
-  const [isLoading, toggleIsLoading] = useState(hasNativeLazyLoadSupport)
+  const [isLoading, toggleIsLoading] = useState(hasNativeLazyLoadSupport())
   const [isLoaded, toggleIsLoaded] = useState(false)
 
   if (!global.GATSBY___IMAGE && !hasShownWarning) {
@@ -91,7 +100,7 @@ export const GatsbyImageHydrator: FunctionComponent<GatsbyImageProps> = function
       ) as HTMLImageElement
 
       // when SSR and native lazyload is supported we'll do nothing ;)
-      if (hasNativeLazyLoadSupport && hasSSRHtml && global.GATSBY___IMAGE) {
+      if (hasNativeLazyLoadSupport() && hasSSRHtml && global.GATSBY___IMAGE) {
         onStartLoad?.({ wasCached: false })
 
         if (hasSSRHtml.complete) {
@@ -140,8 +149,14 @@ export const GatsbyImageHydrator: FunctionComponent<GatsbyImageProps> = function
     if (root.current) {
       const hasSSRHtml = root.current.querySelector(`[data-gatsby-image-ssr]`)
       // On first server hydration do nothing
-      if (hasNativeLazyLoadSupport && hasSSRHtml && !hydrated.current) {
+      if (hasNativeLazyLoadSupport() && hasSSRHtml && !hydrated.current) {
+        hydrated.current = true
         return
+      }
+
+      // When no ssrHtml is found (develop) we should force render instead of hydrate
+      if (!hasSSRHtml) {
+        hydrated.current = true
       }
 
       import(`./lazy-hydrate`).then(({ lazyHydrate }) => {
@@ -180,6 +195,7 @@ export const GatsbyImageHydrator: FunctionComponent<GatsbyImageProps> = function
       style={{
         ...wStyle,
         ...style,
+        backgroundColor,
       }}
       className={`${wClass}${className ? ` ${className}` : ``}`}
       ref={root}