+++ packages/gatsby-plugin-image/src/components/hooks.ts
@@ -11,12 +11,16 @@ import {
 import { Node } from "gatsby"
 import { PlaceholderProps } from "./placeholder"
 import { MainImageProps } from "./main-image"
-import { Layout } from "../utils"
-import { ISharpGatsbyImageData } from "./gatsby-image.browser"
+import type { IGatsbyImageData } from "./gatsby-image.browser"
+import {
+  IGatsbyImageHelperArgs,
+  generateImageData,
+  Layout,
+} from "../image-utils"
 const imageCache = new Set<string>()
 
 // Native lazy-loading support: https://addyosmani.com/blog/lazy-loading/
-export const hasNativeLazyLoadSupport =
+export const hasNativeLazyLoadSupport = (): boolean =>
   typeof HTMLImageElement !== `undefined` &&
   `loading` in HTMLImageElement.prototype
 
@@ -32,12 +36,15 @@ export function hasImageLoaded(cacheKey: string): boolean {
 
 export type FileNode = Node & {
   childImageSharp?: Node & {
-    gatsbyImageData?: ISharpGatsbyImageData
+    gatsbyImageData?: IGatsbyImageData
   }
 }
 
-export const getImage = (file: FileNode): ISharpGatsbyImageData | undefined =>
+export const getImage = (file: FileNode): IGatsbyImageData | undefined =>
   file?.childImageSharp?.gatsbyImageData
+
+export const getSrc = (file: FileNode): string | undefined =>
+  file?.childImageSharp?.gatsbyImageData?.images?.fallback?.src
 
 export function getWrapperProps(
   width: number,
@@ -48,14 +55,13 @@ export function getWrapperProps(
 } {
   const wrapperStyle: CSSProperties = {
     position: `relative`,
+    overflow: `hidden`,
   }
 
   if (layout === `fixed`) {
     wrapperStyle.width = width
     wrapperStyle.height = height
-  }
-
-  if (layout === `constrained`) {
+  } else if (layout === `constrained`) {
     wrapperStyle.display = `inline-block`
   }
 
@@ -66,6 +72,14 @@ export function getWrapperProps(
   }
 }
 
+export function useGatsbyImage({
+  pluginName = `useGatsbyImage`,
+  ...args
+}: IGatsbyImageHelperArgs): IGatsbyImageData {
+  // TODO: use context to get default plugin options and spread them in here
+  return generateImageData({ pluginName, ...args })
+}
+
 export function getMainProps(
   isLoading: boolean,
   isLoaded: boolean,
@@ -73,7 +87,8 @@ export function getMainProps(
   loading?: "eager" | "lazy",
   toggleLoaded?: any,
   cacheKey?: string,
-  ref?: any
+  ref?: any,
+  style: CSSProperties = {}
 ): MainImageProps {
   const onLoad: ReactEventHandler<HTMLImageElement> = function (e) {
     if (isLoaded) {
@@ -101,29 +116,32 @@ export function getMainProps(
     }
   }
 
+  // fallback when it's not configured in gatsby-config.
+  if (!global.GATSBY___IMAGE) {
+    style = {
+      height: `100%`,
+      left: 0,
+      position: `absolute`,
+      top: 0,
+      transform: `translateZ(0)`,
+      transition: `opacity 250ms linear`,
+      width: `100%`,
+      willChange: `opacity`,
+      ...style,
+    }
+  }
+
   const result = {
     ...images,
     loading,
     shouldLoad: isLoading,
     "data-main-image": ``,
     style: {
+      ...style,
       opacity: isLoaded ? 1 : 0,
     },
     onLoad,
     ref,
-  }
-
-  // fallback when it's not configured in gatsby-config.
-  if (!global.GATSBY___IMAGE) {
-    result.style.height = `100%`
-    result.style.left = 0
-    result.style.position = `absolute`
-    result.style.top = 0
-    result.style.transform = `translateZ(0)`
-    result.style.transition = `opacity 250ms linear`
-    result.style.width = `100%`
-    result.style.willChange = `opacity`
-    result.style.objectFit = `cover`
   }
 
   return result
@@ -145,17 +163,26 @@ export function getPlaceholderProps(
   const wrapperStyle: CSSProperties = {}
 
   if (backgroundColor) {
+    wrapperStyle.backgroundColor = backgroundColor
+
     if (layout === `fixed`) {
       wrapperStyle.width = width
       wrapperStyle.height = height
+      wrapperStyle.backgroundColor = backgroundColor
+      wrapperStyle.position = `relative`
+    } else if (layout === `constrained`) {
+      wrapperStyle.position = `absolute`
+      wrapperStyle.top = 0
+      wrapperStyle.left = 0
+      wrapperStyle.bottom = 0
+      wrapperStyle.right = 0
+    } else if (layout === `fluid`) {
+      wrapperStyle.position = `absolute`
+      wrapperStyle.top = 0
+      wrapperStyle.left = 0
+      wrapperStyle.bottom = 0
+      wrapperStyle.right = 0
     }
-
-    if (layout === `constrained`) {
-      wrapperStyle.display = `inline-block`
-    }
-
-    wrapperStyle.backgroundColor = backgroundColor
-    wrapperStyle.position = `relative`
   }
 
   const result: PlaceholderImageAttrs = {