+++ packages/gatsby-transformer-sharp/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.10.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sharp@2.10.0-next.0...gatsby-transformer-sharp@2.10.0-next.1) (2020-12-17)
+
+### Features
+
+- **gatsby-plugin-image:** Make dimensions optional and default to constrained ([#28662](https://github.com/gatsbyjs/gatsby/issues/28662)) ([f24a23d](https://github.com/gatsbyjs/gatsby/commit/f24a23d5b1f24ab33aec7a8ac882dbbccdf7d01e))
+
+# [2.10.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sharp@2.9.0-next.0...gatsby-transformer-sharp@2.10.0-next.0) (2020-12-10)
+
+**Note:** Version bump only for package gatsby-transformer-sharp
+
+# [2.9.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sharp@2.8.0-next.1...gatsby-transformer-sharp@2.9.0-next.0) (2020-11-26)
+
+**Note:** Version bump only for package gatsby-transformer-sharp
+
+# [2.8.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sharp@2.8.0-next.0...gatsby-transformer-sharp@2.8.0-next.1) (2020-11-18)
+
+**Note:** Version bump only for package gatsby-transformer-sharp
+
+# [2.8.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sharp@2.7.0-next.1...gatsby-transformer-sharp@2.8.0-next.0) (2020-11-18)
+
+**Note:** Version bump only for package gatsby-transformer-sharp
+
+# [2.7.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sharp@2.7.0-next.0...gatsby-transformer-sharp@2.7.0-next.1) (2020-11-16)
+
+### Bug Fixes
+
+- **gatsby-transformer-sharp:** create child nodes only for Files ([#27992](https://github.com/gatsbyjs/gatsby/issues/27992)) ([001e045](https://github.com/gatsbyjs/gatsby/commit/001e045c92dfd37a7d573d75ee9d81ae44c8e7a5))
+
 # [2.6.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sharp@2.6.0-next.0...gatsby-transformer-sharp@2.6.0-next.1) (2020-11-10)
 
 ### Features